requirements.txt
astroid==2.3.2
colorama==0.4.1
Django==2.2.6
django-appconf==1.0.3
django-compressor==2.3
django-libsass==0.7
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
isort==4.3.21
Jinja2==2.10.3
lazy-object-proxy==1.4.3
libsass==0.19.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.740
mypy-extensions==0.4.3
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.3
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
wrapt==1.11.2
# rcssmin --install-option="--without-c-extensions"
# django-compressor --upgrade

